Move address attribute definition into io.grpc.
