+++ alts/build.gradle
@@ -57,10 +57,7 @@ import net.ltgt.gradle.errorprone.CheckSeverity
     it.options.errorprone.check("FutureReturnValueIgnored", CheckSeverity.OFF)
 }
 
-javadoc {
-    exclude 'io/grpc/alts/internal/**'
-    exclude 'io/grpc/alts/Internal?*.java'
-}
+javadoc { exclude 'io/grpc/alts/internal/**' }
 
 jar {
     // Must use a different classifier to avoid conflicting with shadowJar
